Convert internal date to Y-m-d format.
Previously, the XML stored dates in <announce> and <revised>
tags in "F d, Y" format (January 2, 1984). This commit
will change the behaviour to the equivalent (and in accordance
with the DTD) format of "Y-m-d" (1984-01-02).
Display of that value in the Emails will be consistent with
the old format, the XSLT can handle the new format fine.

Furthermore, the "count" attribute of <revised> is used
to save revision counts, whereas it was just appended
to the date before.

Display of dates within GLSAmaker will change.