Ingest xml tags without newlines
Ingesting the newlines as written in the XML causes problems if we
ever want to use or display the data with different formatting. By
stripping the newlines, we can use whatever formatting we want when
the data is used. This enables us to fix many of the problems in GLSA
mail formatting by wrapping lines in the template itself.

Signed-off-by: John Helmert III <ajak@gentoo.org>