glsamaker/autoglsa: recovery nicer when we misparse bug summaries
This includes a bit of refactoring of autoglsa.get_max_versions to
allow for easy recovery when parsing of a particular bug fails. Bugs
which fail the "autoglsa" process will be returned as a list of
exceptions for the caller to handle as necessary. Currently, this
means spitting out an error message on the edit_glsa page.

Signed-off-by: John Helmert III <ajak@gentoo.org>