models/glsa: catch another edge case in mail table generation
We were always using the "vulnerable version" objects to find which
package name to output into the table, but that doesn't work if
there's no vulnerable version. Since there's always one package in
either of the vulnerable or unaffected objects, use whichever happens
to provide the string we're looking for.

Since we were (and are) relying on the caller providing us a pair of
queries which includes only one package, raise an exception if that
assumption is broken.

Also add a test case which catches the original issue.

Signed-off-by: John Helmert III <ajak@gentoo.org>